py-backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
python-multipart
pdf2image
# sudo apt-get install poppler-utils